# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenClinic — Electronic Health Record (EHR) & Governance API",
    "description": "Open, modular, and auditable Electronic Health Record (EHR) and Clinical Management API with hybrid access control (RBAC + ACL), multi-tenancy, dynamic capabilities resolution, and enterprise security (Argon2id + in-memory JWT).",
    "version": "1.0.0",
    "contact": {
      "name": "Iniciativa OpenClinic",
      "url": "https://github.com/Iniciativa-OpenClinic/OpenClinic"
    },
    "license": {
      "name": "GNU AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token in Authorization header: `Bearer <token>`"
      }
    },
    "schemas": {
...
      "def-0": {
        "type": "object",
        "description": "Estrutura padronizada de erros da aplicação no padrão RFC 7807",
        "properties": {
          "type": {
            "type": "string",
            "description": "URI identificadora do tipo de erro"
          },
          "title": {
            "type": "string",
            "description": "Resumo legível do erro"
          },
          "status": {
            "type": "integer",
            "description": "Código de status HTTP"
          },
          "detail": {
            "type": "string",
            "description": "Descrição detalhada e traduzida do erro"
          },
          "code": {
            "type": "string",
            "description": "Código unívoco de erro da aplicação (ex: ERR_AUTH_INVALID_CREDENTIALS)"
          },
          "instance": {
            "type": "string",
            "description": "URI do endpoint onde ocorreu o erro"
          },
          "invalid_params": {
            "type": "array",
            "description": "Lista de parâmetros com falha de validação",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "reason": {
                  "type": "string"
                }
              }
            }
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetails"
      }
    }
  },
  "paths": {
...
    "/api/v1/public/config": {
      "get": {
        "summary": "Get Public Application Branding and Settings",
        "tags": [
          "Architecture & Platform"
        ],
        "description": "Returns publicly accessible application branding (appSubtitle, appLogoUrl, appFaviconUrl, locales) from database for unauthenticated bootstrap.",
        "responses": {
          "200": {
            "description": "Public application branding parameters",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Public application branding parameters",
                  "type": "object",
                  "properties": {
                    "appName": {
                      "type": "string"
                    },
                    "appSubtitle": {
                      "type": "string",
                      "nullable": true
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "appLogoUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "appFaviconUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "appDescription": {
                      "type": "string",
                      "nullable": true
                    },
                    "tenantName": {
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string"
                    },
                    "defaultSupportedLocales": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "defaultTimezone": {
                      "type": "string"
                    },
                    "defaultDialingCode": {
                      "type": "string"
                    },
                    "acceptedLoginMethods": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "primaryLoginIdentifier": {
                      "type": "string",
                      "enum": [
                        "CPF",
                        "USERNAME",
                        "EMAIL"
                      ]
                    },
                    "app_name": {
                      "type": "string"
                    },
                    "app_subtitle": {
                      "type": "string",
                      "nullable": true
                    },
                    "app_version": {
                      "type": "string"
                    },
                    "app_logo_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "app_favicon_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "tenant_name": {
                      "type": "string"
                    },
                    "default_locale": {
                      "type": "string"
                    },
                    "supported_locales": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "default_timezone": {
                      "type": "string"
                    },
                    "default_dialing_code": {
                      "type": "string"
                    },
                    "accepted_login_methods": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "primary_login_identifier": {
                      "type": "string",
                      "enum": [
                        "CPF",
                        "USERNAME",
                        "EMAIL"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida ou erro de validação (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado ou token inválido/expirado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado — Permissão ou Role insuficiente (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/arch/public/application": {
      "get": {
        "summary": "Get Public Application Branding and Settings",
        "tags": [
          "Architecture & Platform"
        ],
        "description": "Returns publicly accessible application branding (appSubtitle, appLogoUrl, appFaviconUrl, locales) from database for unauthenticated bootstrap.",
        "responses": {
          "200": {
            "description": "Public application branding parameters",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Public application branding parameters",
                  "type": "object",
                  "properties": {
                    "appName": {
                      "type": "string"
                    },
                    "appSubtitle": {
                      "type": "string",
                      "nullable": true
                    },
                    "appVersion": {
                      "type": "string"
                    },
                    "appLogoUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "appFaviconUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "appDescription": {
                      "type": "string",
                      "nullable": true
                    },
                    "tenantName": {
                      "type": "string"
                    },
                    "defaultLocale": {
                      "type": "string"
                    },
                    "defaultSupportedLocales": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "defaultTimezone": {
                      "type": "string"
                    },
                    "defaultDialingCode": {
                      "type": "string"
                    },
                    "acceptedLoginMethods": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "primaryLoginIdentifier": {
                      "type": "string",
                      "enum": [
                        "CPF",
                        "USERNAME",
                        "EMAIL"
                      ]
                    },
                    "app_name": {
                      "type": "string"
                    },
                    "app_subtitle": {
                      "type": "string",
                      "nullable": true
                    },
                    "app_version": {
                      "type": "string"
                    },
                    "app_logo_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "app_favicon_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "tenant_name": {
                      "type": "string"
                    },
                    "default_locale": {
                      "type": "string"
                    },
                    "supported_locales": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "default_timezone": {
                      "type": "string"
                    },
                    "default_dialing_code": {
                      "type": "string"
                    },
                    "accepted_login_methods": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "primary_login_identifier": {
                      "type": "string",
                      "enum": [
                        "CPF",
                        "USERNAME",
                        "EMAIL"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida ou erro de validação (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado ou token inválido/expirado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado — Permissão ou Role insuficiente (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor (RFC 7807)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
}